+++ CHANGELOG.md
@@ -7,6 +7,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this
 project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.3.0] - 2023-01-28
+### Added
+- `transform` option to `title` plugin.
+
 ## [0.2.0] - 2023-01-03
 ### Added
 - `image` plugin.
@@ -20,5 +24,6 @@ project adheres to [Semantic Versioning](http://semver.org/).
 - `toc` plugin.
 - `title` plugin.
 
+[0.3.0]: https://github.com/lumeland/markdown-plugins/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/lumeland/markdown-plugins/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/lumeland/markdown-plugins/releases/tag/v0.1.0